on continu le nettoyage

+++ src/main/java/org/crossfit/app/web/rest/api/BookingResource.java
@@ -10,6 +10,8 @@ import javax.validation.Valid;
 
 import org.crossfit.app.domain.Booking;
 import org.crossfit.app.repository.BookingRepository;
+import org.crossfit.app.security.SecurityUtils;
+import org.crossfit.app.service.CrossFitBoxSerivce;
 import org.crossfit.app.web.rest.util.HeaderUtil;
 import org.crossfit.app.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
@@ -38,39 +40,8 @@ public class BookingResource {
     @Inject
     private BookingRepository bookingRepository;
 
-    /**
-     * POST  /bookings -> Create a new booking.
-     */
-    @RequestMapping(value = "/bookings",
-            method = RequestMethod.POST,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Booking> create(@Valid @RequestBody Booking booking) throws URISyntaxException {
-        log.debug("REST request to save Booking : {}", booking);
-        if (booking.getId() != null) {
-            return ResponseEntity.badRequest().header("Failure", "A new booking cannot already have an ID").body(null);
-        }
-        Booking result = bookingRepository.save(booking);
-        return ResponseEntity.created(new URI("/api/bookings/" + result.getId()))
-                .headers(HeaderUtil.createEntityCreationAlert("booking", result.getId().toString()))
-                .body(result);
-    }
-
-    /**
-     * PUT  /bookings -> Updates an existing booking.
-     */
-    @RequestMapping(value = "/bookings",
-        method = RequestMethod.PUT,
-        produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Booking> update(@Valid @RequestBody Booking booking) throws URISyntaxException {
-        log.debug("REST request to update Booking : {}", booking);
-        if (booking.getId() == null) {
-            return create(booking);
-        }
-        Booking result = bookingRepository.save(booking);
-        return ResponseEntity.ok()
-                .headers(HeaderUtil.createEntityUpdateAlert("booking", booking.getId().toString()))
-                .body(result);
-    }
+    @Inject
+    private CrossFitBoxSerivce boxService;
 
     /**
      * GET  /bookings -> get all the bookings.
@@ -81,25 +52,11 @@ public class BookingResource {
     public ResponseEntity<List<Booking>> getAll(@RequestParam(value = "page" , required = false) Integer offset,
                                   @RequestParam(value = "per_page", required = false) Integer limit)
         throws URISyntaxException {
-        Page<Booking> page = bookingRepository.findAll(PaginationUtil.generatePageRequest(offset, limit));
+        Page<Booking> page = bookingRepository.findAllByMember(boxService.findCurrentCrossFitBox(), SecurityUtils.getCurrentMember(), PaginationUtil.generatePageRequest(offset, limit));
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/bookings", offset, limit);
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
 
-    /**
-     * GET  /bookings/:id -> get the "id" booking.
-     */
-    @RequestMapping(value = "/bookings/{id}",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Booking> get(@PathVariable Long id) {
-        log.debug("REST request to get Booking : {}", id);
-        return Optional.ofNullable(bookingRepository.findOne(id))
-            .map(booking -> new ResponseEntity<>(
-                booking,
-                HttpStatus.OK))
-            .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
-    }
 
     /**
      * DELETE  /bookings/:id -> delete the "id" booking.
@@ -109,6 +66,11 @@ public class BookingResource {
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> delete(@PathVariable Long id) {
         log.debug("REST request to delete Booking : {}", id);
+		Booking booking = bookingRepository.findOne(id);
+		
+		if(booking == null || !booking.getOwner().equals( SecurityUtils.getCurrentMember())){
+			return ResponseEntity.badRequest().header("Failure", "Vous n'êtes pas le propiétaire de cette réservation").body(null);
+		}
         bookingRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("booking", id.toString())).build();
     }
